Modify: phase2-2. 해당 테스트는 기상청의 API 결과에 너무 종속적이지만 필요한 테스트이기 때문에 SpringBootTest에서 제거하고 패키지도 변경 [#2]
